Using specific commit for EDD License Handler
EDD went ahead and removed all the source code in the repository, with the justification that it was just "example code".
Now the repository only contains a README which demonstrates usage of the EDD-provided license handler.
This change makes the dev-dependency stick to the commit prior to the repo-fuckery made by EDD, until a better solution is found.